backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg[binary]==3.2.12
python-dotenv==1.0.0
pydantic==2.9.2
python-multipart==0.0.6
qrcode[pil]==7.4.2


